docs : change ddl type

+++ src/main/resources/schema.sql
@@ -1,5 +1,5 @@
 CREATE TABLE stadium (
-    id INT AUTO_INCREMENT PRIMARY KEY,
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
     name VARCHAR(100) not null,
     address VARCHAR(100) not null,
     size VARCHAR(50) not null,
